+++ reikna/cluda/api.py
@@ -498,13 +498,16 @@ class Kernel:
         """
         Execute the kernel.
         :py:class:`Array` objects are allowed as arguments.
+        In case of the OpenCL backend, returns a ``pyopencl.Event`` object.
         """
-        self._prepared_call(*args)
+        result = self._prepared_call(*args)
         self._thr._synchronize()
+        return result
 
     def __call__(self, *args, **kwds):
         """
         A shortcut for successive call to :py:meth:`prepare` and :py:meth:`prepared_call`.
+        In case of the OpenCL backend, returns a ``pyopencl.Event`` object.
         """
         if not self._static:
             if 'global_size' in kwds:
@@ -513,7 +516,7 @@ class Kernel:
                 raise TypeError("global_size keyword argument must be set")
             self.prepare(*prep_args, **kwds)
 
-        self.prepared_call(*args)
+        return self.prepared_call(*args)
 
 
 class StaticKernel:
@@ -585,5 +588,6 @@ class StaticKernel:
         """
         Execute the kernel.
         :py:class:`Array` objects are allowed as arguments.
+        In case of the OpenCL backend, returns a ``pyopencl.Event`` object.
         """
-        self._kernel(*args)
+        return self._kernel(*args)